test(components/atom/pinInput): api tab
