+++ components/atom/pinInput/demo/ArticleDefault.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import {Article, H2, Paragraph, Code} from '@s-ui/documentation-library'
-import PinInput from '../src/PinInput'
+import PinInput from '../src'
 import {useState} from 'react'
 
 const ArticleDefault = ({className}) => {
