+++ components/atom/pinInput/demo/ArticleDisabled.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import {Article, H2, Paragraph, RadioButton} from '@s-ui/documentation-library'
-import PinInput from '../src/PinInput'
+import PinInput from '../src'
 import {useState} from 'react'
 
 const ArticleDisabled = ({className}) => {
