+++ components/atom/pinInput/demo/ArticleLength.js
@@ -7,7 +7,7 @@ import {
   Grid,
   Cell
 } from '@s-ui/documentation-library'
-import PinInput from '../src/PinInput'
+import PinInput from '../src'
 import {useState} from 'react'
 
 const ArticleLenght = ({className}) => {
