+++ components/atom/pinInput/demo/ArticleReferenced.js
@@ -8,7 +8,7 @@ import {
   Cell,
   Label
 } from '@s-ui/documentation-library'
-import PinInput from '../src/PinInput'
+import PinInput from '../src'
 import {useState} from 'react'
 
 const ArticleReferenced = ({className}) => {
