+++ components/atom/pinInput/demo/ArticleSizes.js
@@ -6,7 +6,7 @@ import {
   RadioButton,
   RadioButtonGroup
 } from '@s-ui/documentation-library'
-import PinInput from '../src/PinInput'
+import PinInput from '../src'
 import {useState} from 'react'
 import {SIZES} from '../src/config'
